# PlaceCar_2024_Angular/public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PlaceCar API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Users/Login": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Users/NouveauClient": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddClientDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddClientDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddClientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Users/NouveauEmployee": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEmplDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEmplDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddEmplDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Users/NouveauAdmin": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEmplDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEmplDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddEmplDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddClientDto": {
        "type": "object",
        "properties": {
          "perS_Nom": {
            "type": "string",
            "nullable": true
          },
          "perS_Prenom": {
            "type": "string",
            "nullable": true
          },
          "perS_DateNaissance": {
            "type": "string",
            "format": "date-time"
          },
          "perS_Email": {
            "type": "string",
            "nullable": true
          },
          "perS_Password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AddEmplDto": {
        "type": "object",
        "properties": {
          "perS_Nom": {
            "type": "string",
            "nullable": true
          },
          "perS_Prenom": {
            "type": "string",
            "nullable": true
          },
          "perS_DateNaissance": {
            "type": "string",
            "format": "date-time"
          },
          "perS_Email": {
            "type": "string",
            "nullable": true
          },
          "perS_Password": {
            "type": "string",
            "nullable": true
          },
          "parkingId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Encoder votre toke; Exemple : Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}